+++ CHANGELOG.md
@@ -5,10 +5,14 @@ All notable changes to SimpleBroker will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.7.2] - 2025-10-27
+### Added
+- Added .get_meta method on BrokerCore/BrokerDB to provide a readout of the meta table
+- Added additional tests
+
 ## [2.7.1] - 2025-10-24
 ### Changed
 - Updated tests on windows to use unbuffered I/O and longer timeouts
-
 
 ## [2.7.0] - 2025-10-24
 ### Added